server/requirements.txt
Flask
flask_restplus
flask_mongoalchemy
flask_jwt_extended
flask_cors==3.0.7
flask-login